// tooling/types.ts
export type TaskArgsOption = {
    type: "string" | "boolean";
    required?: boolean;
    description?: string;
    default?: string | boolean | string[] | boolean[] | undefined;
    choices?: string[];
};

export type TaskArgsOptions = {
    [key: string]: TaskArgsOption;
};

export type TaskMeta = {
    name: string;
    description: string;
    options?: TaskArgsOptions;
    positionals?: {
        name: string;
        description?: string;
        required?: boolean;
    };
};

export type TaskArgs = {[key: string]: string | string[] | boolean};

export type TaskFunction = (taskArgs: TaskArgs, tooling: Tooling) => Promise<void>;